Adicion Lombok para Lectura y adicion de constructor..
