--- src/main/java/com/fundamental/model/Lectura.java
@@ -1,148 +1,0 @@
-package com.fundamental.model;
-
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-
-/**
- * @author Henry Barrientos
- */
-public class Lectura {
-
-    private Integer lecturaId, estacionId, turnoId;
-    private String creadoPor;
-    private Date creadoEl;
-    private String creadoPersona, modificadoPor, descError  /**, nombrePistero, nombreJefe**/;
-    private Date modificadoEl;
-    private List<LecturaDetalle> lecturaDetalle = new ArrayList();
-    private List<Lecturafinal> lecturafinal = new ArrayList();
-    private Integer empleadoId;
-
-    public Lectura(Integer lecturaId, Integer estacionId, Integer turnoId
-                    , String creadoPor, String creadoPersona/**, String nombrePistero, String nombreJefe**/) {
-        this.lecturaId = lecturaId;
-        this.estacionId = estacionId;
-        this.turnoId = turnoId;
-        this.creadoPor = creadoPor;
-        this.creadoPersona = creadoPersona;
-        //this.nombrePistero = nombrePistero;
-        //this.nombreJefe = nombreJefe;
-    }
-
-    public Lectura() {
-//        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    public Integer getLecturaId() {
-        return lecturaId;
-    }
-
-    public void setLecturaId(Integer lecturaId) {
-        this.lecturaId = lecturaId;
-    }
-
-    public Integer getEstacionId() {
-        return estacionId;
-    }
-
-    public void setEstacionId(Integer estacionId) {
-        this.estacionId = estacionId;
-    }
-
-    public Integer getTurnoId() {
-        return turnoId;
-    }
-
-    public void setTurnoId(Integer turnoId) {
-        this.turnoId = turnoId;
-    }
-
-    public String getCreadoPor() {
-        return creadoPor;
-    }
-
-    public void setCreadoPor(String creadoPor) {
-        this.creadoPor = creadoPor;
-    }
-
-    public Date getCreadoEl() {
-        return creadoEl;
-    }
-
-    public void setCreadoEl(Date creadoEl) {
-        this.creadoEl = creadoEl;
-    }
-
-    public String getCreadoPersona() {
-        return creadoPersona;
-    }
-
-    public void setCreadoPersona(String creadoPersona) {
-        this.creadoPersona = creadoPersona;
-    }
-
-    public String getModificadoPor() {
-        return modificadoPor;
-    }
-
-    public void setModificadoPor(String modificadoPor) {
-        this.modificadoPor = modificadoPor;
-    }
-
-    public Date getModificadoEl() {
-        return modificadoEl;
-    }
-
-    public void setModificadoEl(Date modificadoEl) {
-        this.modificadoEl = modificadoEl;
-    }
-
-    public String getDescError() {
-        return descError;
-    }
-
-    public void setDescError(String descError) {
-        this.descError = descError;
-    }
-
-    public List<LecturaDetalle> getLecturaDetalle() {
-        return lecturaDetalle;
-    }
-
-    public void setLecturaDetalle(List<LecturaDetalle> lecturaDetalle) {
-        this.lecturaDetalle = lecturaDetalle;
-    }
-
-    public List<Lecturafinal> getLecturafinal() {
-        return lecturafinal;
-    }
-
-    public void setLecturafinal(List<Lecturafinal> lecturafinal) {
-        this.lecturafinal = lecturafinal;
-    }
-
-//    public String getNombrePistero() {
-//        return nombrePistero;
-//    }
-//
-//    public void setNombrePistero(String nombrePistero) {
-//        this.nombrePistero = nombrePistero;
-//    }
-//
-//    public String getNombreJefe() {
-//        return nombreJefe;
-//    }
-//
-//    public void setNombreJefe(String nombreJefe) {
-//        this.nombreJefe = nombreJefe;
-//    }
-
-    public Integer getEmpleadoId() {
-        return empleadoId;
-    }
-
-    public void setEmpleadoId(Integer empleadoId) {
-        this.empleadoId = empleadoId;
-    }
-
-}
